+++ Pwn-Power/main/wifi_scan.h
@@ -19,7 +19,7 @@ bool wifi_scan_was_truncated();
 bool wifi_scan_has_new_results();
 void wifi_scan_update_ui_cache_from_record(const scan_record_t *record);
 extern const uint8_t* get_scan_channels(void);
-extern const size_t get_scan_channels_size(void);
+extern size_t get_scan_channels_size(void);
 
 typedef struct {
     uint8_t station_mac[6];